default commit message

+++ menu.sh
@@ -1,19 +1,8 @@
 #!/bin/bash
 
-# This sources all the functions in the ms_mac_functions.sh file
-# These will be used to perform the actions in the menu
-# script_dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" &>/dev/null && pwd)"
-# echo "Script directory: $script_dir"
-
-# Initial full path
-
-
-
-# Navigate up two levels
-# VSCODE_PATH=$(dirname $(dirname "$script_dir"))
-
-# Print the result
 # echo "VSCode path: $VSCODE_PATH"
+
+clear
 
 #Key folders
 CODE_FOLDER_PATH="${HOME}/code"
@@ -23,31 +12,54 @@ PRODUCT_TOOLS_PATH="${CODE_FOLDER_PATH}/product-tools"
 CODE_ADMIN_PATH="${CODE_FOLDER_PATH}/code-admin"
 
 #Color Variables for text
+BLACK='\033[0;30m'
 RED='\033[0;31m'
+RED_U='\033[4;31m'
+RED_BLINK='\033[5;31m'
 GREEN='\033[0;32m'
-BLUE='\033[0;31m'
+GREEN_BLINK='\033[5;32m'
+YELLOW='\033[0;33m'
+YELLOW_BOLD='\033[1;33m'
+PURPLE='\033[1;34m'
+PURPLE_U='\033[4;34m'
+PURPLE_BLINK='\033[5;34m'
+PINK='\033[0;35m'
+PINK_U='\033[4;35m'
+PINK_BLINK='\033[5;35m'
+LIGHTBLUE='\033[0;36m'
+LIGHTBLUE_BOLD='\033[1;36m'
+GRAY='\033[0;37m'
+ORANGE='\033[1;91m'
+BLUE='\033[1;94m'
+CYAN='\033[1;96m'
+WHITE='\033[1;97m'
+MAGENTA='\033[1;95m'
+BOLD='\033[1m'
+UNDERLINE='\033[4m'
+BLINK='\033[5m'
+
 NC='\033[0m' # No Color
 
 # Function to display the menu
 show_menu() {
-    echo -e "\n\nHello! What would you like to do?"
+    echo -e "\n\n${LIGHTBLUE_BOLD}Hello! What would you like to do?"
     echo -e "Please choose one of the following options (enter the number):\n"
-    echo -e "|-------------------------------------------------------------|\n"
+    echo -e "|-------------------------------------------------------------|${NC}\n"
 
-    echo -e "${BLUE}0) Commit the current repository${NC}\n"
-    echo -e "1) Commit and Push All: VSCODE .JBI COMMUNIFY\n"
-    echo -e "${BLUE}2) Start / Restart Stable Tools${NC}\n"
-    echo -e "3) Start Doccano\n"
-    # echo -e "4) Update .zshrc\n"
-    # echo -e "5) Edit default .zshrc file\n"
-    # echo -e "6) Edit default .gitignore file\n"
-    echo -e "7) Restart Postgres 14\n"
-    # echo -e "8) launch e2aserver\n"
-    # echo -e "9) Launch stable\n"
-    echo -e "${BLUE}10) Create/replace .jbi symbolic links${NC}\n"
-    echo -e "101) Generate an Encryption Key\n"
-    echo -e "${BLUE}1977) Deinitialize vscode${NC}\n"
-    echo -e "2008) DISCARD all changes and REPLACE with latest version from Azure Devops\n"
+    echo -e "${GREEN}0) Commit the current repository${NC}\n"
+    echo -e "${YELLOW}1) Commit and Push All: VSCODE .JBI COMMUNIFY${NC}\n"
+    echo -e "${PINK}2) Start / Restart Stable Tools${NC}\n"
+    echo -e "${LIGHTBLUE}3) Start Doccano${NC}\n"
+    echo -e "${YELLOW}4) Launch Communify${NC}\n"
+    echo -e "${PURPLE}5) Display Color Palette${NC}\n"
+    echo -e "${GREEN}6) Grant terminal access to iCloud Drive${NC} \n"
+    echo -e "${YELLOW}7) Restart Postgres 14${NC}\n"
+    echo -e "${GREEN}8) Start / Restart LLMs and Chat Apps ${NC}\n"
+    echo -e "${ORANGE}9) Shut Down LLMs and Chat Apps ${NC}\n"
+    echo -e "${RED_U}10) Create/replace .jbi symbolic links${NC}\n"
+    echo -e "${RED_U}101) Generate an Encryption Key${NC}\n"
+    echo -e "${RED_U}1977) Deinitialize vscode${NC}\n"
+    echo -e "${RED_U}2008) DISCARD all changes and REPLACE with latest version from Azure Devops${NC}\n"
 
 }
 
@@ -102,37 +114,39 @@ read_choice() {
         ;;
 
     2)
-        echo "Update and Start Product Tools"
+        echo "${RED_U}Shutdown, Delete, Pull, Reinstall and Start Stable Tools${NC}"
+        if confirm "Do you want to proceed?"; then
+            # Check if app.py is running
+            APP_PY_PID=$(pgrep -f "streamlit run 000_Communify_Home")
 
-        # Check if app.py is running
-        APP_PY_PID=$(pgrep -f "streamlit run 000_Communify_Home")
+            if [ -n "$APP_PY_PID" ]; then
+                echo "app.py is running. Stopping it..."
+                kill $APP_PY_PID
+            fi
 
-        if [ -n "$APP_PY_PID" ]; then
-            echo "app.py is running. Stopping it..."
-            kill $APP_PY_PID
+            # Reinstall the Environment
+            STABLE_DIR="$HOME/code/.stable"
+            rm -rf $STABLE_DIR
+            mkdir -p $STABLE_DIR
+
+            cd $STABLE_DIR
+
+            git clone "https://michael:$AZURE_DEVOPS_PAT@dev.azure.com/$AZURE_DEVOPS_ORG/product-development/_git/communify"
+            cd communify
+            ./env_setup.sh
+
+            cd $STABLE_DIR
+
+            git clone "https://michael:$AZURE_DEVOPS_PAT@dev.azure.com/$AZURE_DEVOPS_ORG/product-development/_git/product-tools"
+            cd product-tools
+            ./env_setup.sh
+
+            # Relaunch app.py with streamlit
+            echo "Relaunching app.py with streamlit..."
+            source "$HOME/code/.stable/communify/communify_venv/bin/activate" && streamlit run "$HOME/code/.stable/communify/000_Communify_Home.py"
+        else
+            echo "Returning to menu."
         fi
-
-        # Reinstall the Environment
-        STABLE_DIR="$HOME/code/.stable"
-        rm -rf $STABLE_DIR
-        mkdir -p $STABLE_DIR
-
-        cd $STABLE_DIR
-
-        git clone "https://michael:$AZURE_DEVOPS_PAT@dev.azure.com/$AZURE_DEVOPS_ORG/product-development/_git/communify"
-        cd communify
-        ./env_setup.sh
-
-        cd $STABLE_DIR
-
-        git clone "https://michael:$AZURE_DEVOPS_PAT@dev.azure.com/$AZURE_DEVOPS_ORG/product-development/_git/product-tools"
-        cd product-tools
-        ./env_setup.sh
-
-        # Relaunch app.py with streamlit
-        echo "Relaunching app.py with streamlit..."
-        source "$HOME/code/.stable/communify/communify_venv/bin/activate" && streamlit run "$HOME/code/.stable/communify/000_Communify_Home.py"
-
         ;;
 
     3)
@@ -161,11 +175,32 @@ read_choice() {
         source "$HOME/code/communify/communify_venv/bin/activate" && streamlit run "$HOME/code/communify/000_Communify_Home.py"
         ;;
     5)
-        echo "EMPTY"
-        echo "EMPTY"
+
+        echo "Examples of Font Colors in Bash"
+        echo -e "${BLACK}The Quick Brown Fox Jumped Over the Lazy Dog ${NC}"
+        echo -e "${RED}The Quick Brown Fox Jumped Over the Lazy Dog ${NC}"
+        echo -e "${RED_U}The Quick Brown Fox Jumped Over the Lazy Dog ${NC}"
+        echo -e "${RED_BLINK}The Quick Brown Fox Jumped Over the Lazy Dog ${NC}"
+        echo -e "${GREEN}The Quick Brown Fox Jumped Over the Lazy Dog ${NC}"
+        echo -e "${GREEN_BLINK}The Quick Brown Fox Jumped Over the Lazy Dog ${NC}"
+        echo -e "${YELLOW}The Quick Brown Fox Jumped Over the Lazy Dog ${NC}"
+        echo -e "${PURPLE}The Quick Brown Fox Jumped Over the Lazy Dog ${NC}"
+        echo -e "${PURPLE_U}The Quick Brown Fox Jumped Over the Lazy Dog ${NC}"
+        echo -e "${PURPLE_BLINK}The Quick Brown Fox Jumped Over the Lazy Dog ${NC}"
+        echo -e "${PINK}The Quick Brown Fox Jumped Over the Lazy Dog ${NC}"
+        echo -e "${PINK_U}The Quick Brown Fox Jumped Over the Lazy Dog ${NC}"
+        echo -e "${PINK_BLINK}The Quick Brown Fox Jumped Over the Lazy Dog ${NC}"
+        echo -e "${LIGHTBLUE}The Quick Brown Fox Jumped Over the Lazy Dog ${NC}"
+
         ;;
     6)
-        echo "EMPTY"
+        echo -e "${LIGHTBLUE} Grant Terminal access to iCloud Drive${NC}"
+        echo -e "import os \nprint(os.path.expanduser('~/Library/Mobile Documents/com~apple~CloudDocs')) " >${HOME}/Library/Mobile\ Documents/com~apple~CloudDocs/terminal_access.py
+        cd ${HOME}/Library/Mobile\ Documents/com~apple~CloudDocs
+        python3 ${HOME}/Library/Mobile\ Documents/com~apple~CloudDocs/terminal_access.py
+        echo -e "${GREEN}You should now have access to iCloud Drive in the terminal${NC}"
+        exit 0
+
         ;;
     7)
         echo "Restarting Postgres 14"
@@ -173,21 +208,15 @@ read_choice() {
         echo "Postgres 14 restarted"
         ;;
     8)
-        echo "EMPTY"
-        # # if e2aserver is running, stop it
-        # if docker ps | grep -q e2aserver; then
-        #     echo "e2aserver is running. Stopping it now."
-        #     docker stop e2aserver
-        #     docker rm e2aserver
-        # fi
-        # cd /Volumes/code/vscode/docker/e2aserver/
-        # docker build -t e2aserver .
-        # docker run --name e2aserver -p 5678:5678 -p 8000:8000 --rm -v $(pwd):/app e2aserver
-        # exit 0
+        echo "Start / Restart LLMs and Chat Apps"
+        cd ${COMMUNIFY_FOLDER_PATH} && llm_launch.sh
+        exit 0
         ;;
 
     9)
-        echo "EMPTY"
+        echo "Shut Down LLMs and Chat Apps"
+        cd ${COMMUNIFY_FOLDER_PATH} && llm_launch.sh stop
+        exit 0
         ;;
 
     10)
@@ -218,17 +247,18 @@ read_choice() {
         ;;
 
     2008)
-        echo -e "#### WARNING ####"
-        echo -e "#### WARNING ####\n"
+        echo -e "${RED_U}#### WARNING ####"
+        echo -e "${RED_BLINK}#### WARNING ####\n${NC}"
         echo -e "You are about to DESTROY all changes and"
         echo -e "replace them with the latest version in:"
         echo -e "$pwd\n"
         echo -e "#### THIS CANNOT BE UNDONE ####"
         if confirm "Do you want to proceed?"; then
-            git checkout -- .
-            git reset --hard
-            git fetch origin
-            git reset --hard origin/main
+            echo "Discarding all changes and replacing with latest version from Azure Devops"
+            # git checkout -- .
+            # git reset --hard
+            # git fetch origin
+            # git reset --hard origin/main
         else
             echo "Returning to menu."
         fi
@@ -246,11 +276,11 @@ read_choice() {
 # Function to ask for confirmation
 confirm() {
     while true; do
-        read -p "$1 (y/n): " yn
+        read -p "$1 ( ${PURPLE}Yy or ${PINK_U}Nn ${NC}): " yn
         case $yn in
-        [Yy]*) return 0 ;;                    # User responded yes
-        [Nn]*) return 1 ;;                    # User responded no
-        *) echo "Please answer yes or no." ;; # Invalid response
+        [Yy]*) return 0 ;;                                                     # User responded yes
+        [Nn]*) return 1 ;;                                                     # User responded no
+        *) echo "${RED_BLINK}Please answer Y for 'yes' or N for 'no'.${NC}" ;; # Invalid response
         esac
     done
 }
